Remove pages directory
